refactor: Phase 2 method extraction from CHtmlTables.cs
Extract 5 methods and their private helpers into dedicated classes:
- AddSrvSummaryTable + MapInfraTypeName -> CInfraChipSection
- AddMissingJobsTable -> CMissingJobsTable
- AddBkpSrvTable + AddBackupServerDetails + WriteTupleListToHtml + ConfigDbTable + AddDbCoresRam -> CBackupServerSection
- AddSecSummaryTable + SecurityGridItem -> CSecuritySummarySection
- AddMultiRoleTable -> CMultiRoleTable

CHtmlTables.cs reduced from 2,074 to 1,625 lines (-449 lines).
Original public methods retained as thin delegation wrappers.
