+++ vHC/HC_Reporting/Functions/Reporting/Html/VBR/VbrTables/CHtmlTables.cs
@@ -400,225 +400,10 @@ namespace VeeamHealthCheck.Html.VBR
             return $"<span class=\"badge badge-{variant}\">{System.Net.WebUtility.HtmlEncode(text)}</span>";
         }
 
-        /// <summary>
-        /// Renders a security feature as a grid item with a colored status dot.
-        /// </summary>
-        private static string SecurityGridItem(string label, bool enabled)
-        {
-            string dotColor = enabled ? "green" : "red";
-            return $"<div class=\"security-item\">" +
-                   $"<div class=\"status-dot {dotColor}\"></div>" +
-                   $"<div class=\"label\">{System.Net.WebUtility.HtmlEncode(label)}</div>" +
-                   "</div>";
-        }
-
-        private static string WriteTupleListToHtml(List<Tuple<string, string>> list)
-        {
-            string headers = string.Empty;
-            string data = string.Empty;
-            string s = string.Empty;
-            foreach (var table in list)
-            {
-                headers += table.Item1;
-                data += table.Item2;
-            }
-
-            s += "<table class=\"content-table\"><thead><tr>";
-            s += headers;
-            s += "</tr></thead><tbody><tr>";
-            s += data;
-            s += "</tr></tbody></table>";
-
-            return s;
-        }
-
-        private static string AddBackupServerDetails(BackupServer b)
-        {
-            CVbrServerTable t = new(b);
-            List<Tuple<string, string>> vbrServerData = new(); // t.VbrFullTables();
-
-            if (!CGlobals.RunSecReport)
-            {
-                vbrServerData = t.VbrFullTables();
-            }
-            else
-            {
-                vbrServerData = t.VbrSecurityTables();
-            }
-
-            return WriteTupleListToHtml(vbrServerData);
-        }
-
-        private string ConfigDbTable(BackupServer b)
-        {
-            string s = string.Empty;
-            s += this.form.Subsection("Config Database");
-            s += "<table class=\"content-table\">";
-            s += "<thead><tr>";
-
-            // config DB Table
-            s += this.form.TableHeader("DataBase Type", "MS SQL or PostgreSQL");
-            s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblSqlLocal, VbrLocalizationHelper.BstSqlLocTT);
-            s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblSqlName, VbrLocalizationHelper.BstSqlNameTT);
-            if (b.DbType == CGlobals.SqlTypeName)
-            {
-                s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblSqlVersion, VbrLocalizationHelper.BstSqlVerTT);
-                s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblSqlEdition, VbrLocalizationHelper.BstSqlEdTT);
-            }
-
-            if (b.DbType == CGlobals.SqlTypeName && b.IsLocal == false)
-            {
-                s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblSqlCores, VbrLocalizationHelper.BstSqlCpuTT);
-                s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblSqlRam, VbrLocalizationHelper.BstSqlRamTT);
-            }
-
-            s += this.form.TableHeaderEnd();
-
-            // CDataFormer cd = new(true);
-            try
-            {
-                s += this.form.TableBodyStart();
-                s += "<tr>";
-                s += this.form.TableData(b.DbType, string.Empty);
-                if (b.IsLocal)
-                {
-                    s += this.form.TableData(this.form.True, string.Empty);
-                }
-                else
-                {
-                    s += this.form.TableData(this.form.False, string.Empty);
-                }
-
-                s += this.form.TableData(b.DbHostName, string.Empty);
-                if (b.DbType == CGlobals.SqlTypeName)
-                {
-                    s += this.form.TableData(b.DbVersion, string.Empty);
-                    s += this.form.TableData(b.Edition, string.Empty);
-                }
-
-                if (b.DbType == CGlobals.SqlTypeName && b.IsLocal == false)
-                {
-                    s += this.AddDbCoresRam(b);
-                }
-
-                s += "</tr></tbody></table>";
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Failed to add backup server table. Error:");
-                this.log.Error("\t" + e.Message);
-
-                // return "";
-            }
-
-            return s;
-        }
-
-        private string AddDbCoresRam(BackupServer b)
-        {
-            string s = string.Empty;
-            string dbCoresToolTip = "CPU Cores detected on SQL. 0 indicates SQL is local to VBR or there was an error in collection.";
-            string dbRamToolTip = "RAM detected on SQL. 0 indicates SQL is local to VBR or there was an error in collection.";
-            if (b.DbCores == 0)
-            {
-                s += this.form.TableData("<span class=\"text-muted\">&mdash;</span>", dbCoresToolTip);
-            }
-            else
-            {
-                s += this.form.TableData(b.DbCores.ToString(), dbCoresToolTip);
-            }
-
-            if (b.DbRAM == 0)
-            {
-                s += this.form.TableData("<span class=\"text-muted\">&mdash;</span>", dbRamToolTip);
-            }
-            else
-            {
-                s += this.form.TableData(b.DbRAM.ToString(), dbRamToolTip);
-            }
-
-            return s;
-        }
-
         public string AddBkpSrvTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButtonNoTable("vbrserver", VbrLocalizationHelper.BkpSrvTblHead, string.Empty);
-            string summary = this.sum.SetVbrSummary();
-
-            // CDataFormer cd = new(true);
-            BackupServer b = this.df.BackupServerInfoToXml(scrub);
-            if (String.IsNullOrEmpty(b.Version))
-            {
-                b.Version = CGlobals.VBRFULLVERSION;
-            }
-
-            s += AddBackupServerDetails(b);
-
-            s += this.form.Subsection("Config Backup");
-            s += "<table class=\"content-table\">";
-            s += "<thead><tr>";
-            s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblCfgEnabled, VbrLocalizationHelper.BstCfgEnabledTT);
-            s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblCfgLastRes, VbrLocalizationHelper.BstCfgLastResTT);
-            s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblCfgEncrypt, VbrLocalizationHelper.BstCfgEncTT);
-            s += this.form.TableHeader(VbrLocalizationHelper.BkpSrvTblTarget, VbrLocalizationHelper.BstCfgTarTT);
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-            s += "<tr>";
-            if (b.ConfigBackupEnabled)
-            {
-                s += this.form.TableData(this.form.True, string.Empty);
-            }
-            else
-            {
-                s += this.form.TableData(this.form.False, string.Empty);
-            }
-
-            s += this.form.TableData(Badge(b.ConfigBackupLastResult), string.Empty);
-            if (b.ConfigBackupEncryption)
-            {
-                s += this.form.TableData(this.form.True, string.Empty);
-            }
-            else
-            {
-                s += this.form.TableData(this.form.False, string.Empty);
-            }
-
-            s += this.form.TableData(b.ConfigBackupTarget, string.Empty);
-            s += "</tr></tbody></table>";
-
-            // s += _form.LineBreak();
-            s += this.ConfigDbTable(b);
-
-            // if (CGlobals.RunSecReport)
-            //    s += InstalledAppsTable();
-            s += this.form.SectionEndNoTable(summary);
-
-            // JSON section capture (backup server core info)
-            try
-            {
-                List<string> headers = new() { "Name", "Version", "DbType", "DbHost", "ConfigBackupEnabled", "ConfigBackupLastResult", "ConfigBackupEncryption", "ConfigBackupTarget" };
-                List<List<string>> rows = new()
-                {
-                    new List<string>
-                    {
-                        b.Name,
-                        b.Version,
-                        b.DbType,
-                        b.DbHostName,
-                        b.ConfigBackupEnabled ? "True" : "False",
-                        b.ConfigBackupLastResult,
-                        b.ConfigBackupEncryption ? "True" : "False",
-                        b.ConfigBackupTarget,
-                    },
-                };
-                SetSection("backupServer", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture backupServer JSON section: " + ex.Message);
-            }
-
-            return s;
+            var section = new CBackupServerSection();
+            return section.Render(scrub);
         }
 
         private string InstalledAppsTable()
@@ -640,77 +425,8 @@ namespace VeeamHealthCheck.Html.VBR
 
         public string AddSecSummaryTable(bool scrub)
         {
-            CGlobals.Scrub = scrub;
-
-            string s = this.form.SectionStartWithButtonNoTable("secsummary", VbrLocalizationHelper.SSTitle, string.Empty);
-            string summary = this.sum.SecSum();
-
-            try
-            {
-                CSecuritySummaryTable t = this.df.SecSummary();
-
-                s += "<div class=\"security-grid\">";
-                s += SecurityGridItem("Immutability", t.ImmutabilityEnabled);
-                s += SecurityGridItem("Traffic Encryption", t.TrafficEncrptionEnabled);
-                s += SecurityGridItem("Backup File Encryption", t.BackupFileEncrptionEnabled);
-                s += SecurityGridItem("Config Backup Encryption", t.ConfigBackupEncrptionEnabled);
-                s += SecurityGridItem("MFA Enabled", t.MFAEnabled);
-                s += "</div>";
-            }
-
-            catch (Exception e)
-            {
-                this.log.Error("Security Summary Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            if (CGlobals.VBRMAJORVERSION > 11)
-            {
-                // add malware table
-                try
-                {
-                    var malware = new CMalwareTable();
-                    s += malware.MalwareSettingsTable();
-
-                    s += malware.MalwareExclusionsTable();
-                    s += malware.MalwareEventsTable();
-                    s += malware.MalwareInfectedObjectsTable();
-                }
-                catch (Exception e)
-                {
-                    this.log.Error("Malware Settings Data import failed. ERROR:");
-                    this.log.Error("\t" + e.Message);
-                }
-
-                // Compliance tables extracted to separate sections
-            }
-
-            s += this.form.SectionEndNoTable(summary);
-
-            // JSON section capture security summary
-            try
-            {
-                var t = this.df.SecSummary();
-                List<string> headers = new() { "ImmutabilityEnabled", "TrafficEncryptionEnabled", "BackupFileEncryptionEnabled", "ConfigBackupEncryptionEnabled", "MFAEnabled" };
-                List<List<string>> rows = new()
-                {
-                    new List<string>
-                    {
-                        t.ImmutabilityEnabled ? "True" : "False",
-                        t.TrafficEncrptionEnabled ? "True" : "False",
-                        t.BackupFileEncrptionEnabled ? "True" : "False",
-                        t.ConfigBackupEncrptionEnabled ? "True" : "False",
-                        t.MFAEnabled ? "True" : "False"
-                    },
-                };
-                SetSection("securitySummary", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture securitySummary JSON section: " + ex.Message);
-            }
-
-            return s;
+            var section = new CSecuritySummarySection();
+            return section.Render(scrub);
         }
 
         public string AddComplianceSummaryTable()
@@ -810,56 +526,9 @@ namespace VeeamHealthCheck.Html.VBR
 
         public string AddSrvSummaryTable(bool scrub)
         {
-            string summary = this.sum.SrvSum();
-
-            string s = this.form.SectionStartWithButtonNoTable("serversummary", "Infrastructure Types", "srvSummaryButton");
-
-            try
-            {
-                Dictionary<string, int> list = this.df.ServerSummaryToXml();
-
-                s += "<div class=\"infra-grid\">";
-                foreach (var d in list)
-                {
-                    string infraName = MapInfraTypeName(d.Key);
-                    s += $"<span class=\"infra-chip\"><span class=\"infra-chip-count\">{d.Value}</span>{infraName}</span>";
-                }
-                s += "</div>";
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Server Summary Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEndNoTable(summary);
-
-            // JSON capture server summary
-            try
-            {
-                var list = this.df.ServerSummaryToXml();
-                List<string> headers = new() { "ServerType", "Count" };
-                List<List<string>> rows = list.Select(d => new List<string> { d.Key, d.Value.ToString() }).ToList();
-                SetSection("serverSummary", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture serverSummary JSON section: " + ex.Message);
-            }
-
-            return s;
+            var section = new CInfraChipSection();
+            return section.Render(scrub);
         }
-
-        private static string MapInfraTypeName(string type) => type switch
-        {
-            "WinServer" or "Windows" => "Windows Server",
-            "Linux" => "Linux",
-            "HyperV" => "Hyper-V",
-            "vSphere" or "ESXi" => "VMware vSphere",
-            "AHV" => "Nutanix AHV",
-            "Cloud" => "Cloud",
-            _ => type
-        };
 
         public string AddJobSummaryTable(bool scrub)
         {
@@ -870,65 +539,8 @@ namespace VeeamHealthCheck.Html.VBR
 
         public string AddMissingJobsTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("missingjobs", VbrLocalizationHelper.NpTitle, VbrLocalizationHelper.NpButton);
-
-            string summary = this.sum.MissingJobsSUmmary();
-
-            s += this.form.TableHeaderLeftAligned(VbrLocalizationHelper.JobSum0, string.Empty);
-
-                // _form.TableHeader("Count", "Total detected of this type")
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-
-            // CDataFormer cd = new(true);
-            try
-            {
-                // List<string> list = _df.ParseNonProtectedTypes();
-                CJobSummaryTable st = new();
-                Dictionary<string, int> types = st.JobSummaryTable();
-
-                foreach (var t in types)
-                {
-                    if (t.Value == 0)
-                    {
-                        s += "<tr>";
-                        s += this.form.TableDataLeftAligned(t.Key, string.Empty);
-                        s += "</tr>";
-                    }
-                }
-
-                // for (int i = 0; i < list.Count(); i++)
-                // {
-                //    s += "<tr>";
-
-                // s += _form.TableData(list[i], "");
-                //    s += "</tr>";
-
-                // }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Missing Jobs Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON missing jobs
-            try
-            {
-                CJobSummaryTable st = new();
-                Dictionary<string, int> types = st.JobSummaryTable();
-                List<string> headers = new() { "MissingJobType" };
-                List<List<string>> rows = types.Where(t => t.Value == 0).Select(t => new List<string> { t.Key }).ToList();
-                SetSection("missingJobs", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture missingJobs JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CMissingJobsTable();
+            return table.Render(scrub);
         }
 
         public string AddProtectedWorkLoadsTable(bool scrub)
@@ -1240,68 +852,8 @@ namespace VeeamHealthCheck.Html.VBR
 
         public string AddMultiRoleTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("proxies", VbrLocalizationHelper.PrxTitle, VbrLocalizationHelper.PrxBtn);
-            string summary = this.sum.Proxies();
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx0, VbrLocalizationHelper.Prx0TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx1, VbrLocalizationHelper.Prx1TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx2, VbrLocalizationHelper.Prx2TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx3, VbrLocalizationHelper.Prx3TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx4, VbrLocalizationHelper.Prx4TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx5, VbrLocalizationHelper.Prx5TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx6, VbrLocalizationHelper.Prx6TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx7, VbrLocalizationHelper.Prx7TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx8, VbrLocalizationHelper.Prx8TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx9, VbrLocalizationHelper.Prx9TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx10, VbrLocalizationHelper.Prx10TT);
-            s += this.form.TableHeader(VbrLocalizationHelper.Prx11, VbrLocalizationHelper.Prx11TT);
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-            try
-            {
-                List<string[]> list = this.df.ProxyXmlFromCsv(scrub);
-
-                foreach (var d in list)
-                {
-                    s += "<tr>";
-                    if (scrub)
-                    {
-                        s += this.form.TableData(this.scrub.ScrubItem(d[0], ScrubItemType.Server), string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(d[0], string.Empty);
-                    }
-
-                    s += this.form.TableData(d[1], string.Empty);
-                    s += this.form.TableData(d[2], string.Empty);
-                    s += this.form.TableData(d[3], string.Empty);
-                    s += this.form.TableData(d[4], string.Empty);
-                    s += this.form.TableData(d[5], string.Empty);
-                    s += this.form.TableData(d[6], string.Empty);
-                    s += this.form.TableData(d[7], string.Empty);
-                    s += this.form.TableData(d[8], string.Empty);
-                    s += this.form.TableData(d[9], string.Empty);
-                    if (scrub)
-                    {
-                        s += this.form.TableData(this.scrub.ScrubItem(d[10], ScrubItemType.Server), string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(d[10], string.Empty);
-                    }
-
-                    s += this.form.TableData(d[11], string.Empty);
-                    s += "</tr>";
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("PROXY Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-            return s;
+            var table = new CMultiRoleTable();
+            return table.Render(scrub);
         }
 
         public string AddSobrTable(bool scrub)
